Combining setup scripts
